Use the PID to create the MODS identifier.
UMDMtoMODS.xsl expects the collection/item/{pid,descMeta} structure. It also creates a mods:identifier of type "modsIdentifier" (since that is the type that the MODS_xml_to_rdf.xsl expects) by appending the PID to the end of the http://fedora.lib.umd.edu/fedora/get/ URL.

In the MODS_xml_to_rdf.xsl, don't look for the mods:identifier in the root element, but look relative to the element currently being processed. This allows multiple resources per file.