Utility scripts for getting sample RDF.
getsamplerdf fetches a sample of UMDM from the current Solr index, and runs it through a series of transformations to finally create a Turtle file with triples from DC and MODS RDF with no blank nodes that is suitable for feeding to the umd-fcrepo4-client test DataLoader.

rmbnodes converts a Turtle RDF file to N-Triples, removes all of the triples containing blank nodes, and converts it back to Turtle syntax with the DC and MODS RDF prefixes defined.